+++ MARCELO_PILUSKI_POZZOBON/PrimeiroBI/listaSeis/src/Utils.java
