+++ MARCELO_PILUSKI_POZZOBON/PrimeiroBI/listaSete/src/Utils.java
