+++ MARCELO_PILUSKI_POZZOBON/PrimeiroBI/prova/enunciado.md
@@ -1,0 +1,9 @@
+Prova 03 - Ana Castela, uma coordenadora de eventos, precisa de um sistema para gerenciar eventos e os participantes dela
+
+* Cadastro de cliente 1,5-pt #feito#
+* Cadastro de evento 1,5-pt #feito#
+* Listar eventos 0,5-pt #feito#
+* Comprar ingresso do evento 1,0-pt #feito#
+* Utilizar ingresso em evento 0,5-pt #feito#
+
+* Utilização de conceitos de POO (separação em classes e encapsulamento) *feito meio*
