+++ MARCELO_PILUSKI_POZZOBON/PrimeiroBI/prova/src/Utils.java
@@ -1,0 +1,9 @@
+package MARCELO_PILUSKI_POZZOBON.prova.src;
+
+import java.util.Date;
+
+public class Utils {
+    public static boolean isDataMaior(Date dataUm, Date dataDois) {
+        return dataUm.getTime() > dataDois.getTime();
+    }
+}
