--- MARCELO_PILUSKI_POZZOBON/prova/src/Utils.java
@@ -1,9 +1,0 @@
-package MARCELO_PILUSKI_POZZOBON.prova.src;
-
-import java.util.Date;
-
-public class Utils {
-    public static boolean isDataMaior(Date dataUm, Date dataDois) {
-        return dataUm.getTime() > dataDois.getTime();
-    }
-}
